UART working, QSPI kinda
UART is fully functional - turns out I was just using the wrong baud rate. Additionally, you CAN do UART through the micro usb port and GPIO pins simultaneously! As for QSPI, it's working with the internal flash but not the extrnal one. Something is better than nothing! Although the code in qspi0 might not be the code that worked with internal flash - will need to confirm and update.