Still not very good
Not good at all is a better description... The geiger counter sucks. The display lags upt to 10 seconds behind the actual counter and it's currently not writing to fram. While I think I could have gotten it to work nicely, I underestimated the amount of time it would take to get the display to work and the fram to work in a productive way. Maybe these are just excuses but the display took forever because the existing driver for it did not actually exist in the nrf sdk, so I had to figure out how to manually install a different sdk and get that to work wth a fresh west workspace. The fram wasn't difficult, but it still took a while to write and test everything. The endianness of the data coming back from the fram threw me for a bit of a loop. Sigh.