Custom character progress
Successfully and intentionally added a custom character (a little smiley face) to the display. Still have some figuring out of things to do, but progress has been made! (Also patting myself on the back for pushing a very intermediate stage of the code for once and not forgetting/procrastinating until it's nearly done.)