Debugging/polishing
Found a bug in the driver: The address counter wasn't getting set back to DDRAM after a write to CGRAM for writing a custom character, so the first byte written to the display would actually go to CGRAM instead of being displayed. Fixed that by having the cursor set to the appropriate position at the beginning of the auxdisplay_ac780s_write function. Also polished/cleaned the code a little.