Testing runs with mocked function
It doesn't really do much but I guess I learned how to use FFF amid a struggle of attempting to mock inline functions. (And I of course learned that inline functions can't be mocked. I think I mentioned that in the previous commit, but whatever. Also adding an image of what it looks like when building/running the test natively.