Forgot main again
Forgot to push updates to main.c again. I think this version works, like it prints all the values of sensor to the console, but the version of the driver in the repo at the time of this commit likely (certainly) does not work. Would have made a branch but I'm the only one working in this repository so whatever. (I may regret those words later)