SGP30 Driver Working
Everything save for the feature set and soft reset commands should be good to go. Not sure how I will fit those in due to the limited number of functions exposed to the user of the driver, but hopefully they can be included somehow! Code could use some cleaning up.