MCP9808 Thermometer Driver
Making another driver because I really want to contribute something. This one is not yet in the repo, is not compatible with anything already in the repo, and is not obsolete. I can't think of anything else that would perclude this from being a valid contribution. (other than the fact that it's not finished yet. Still gotta get the alerts and shutdown mode working. Also need to debug the conversion from raw value to int32.